Detail view

+++ app/controllers/orders_controller.rb
@@ -1,4 +1,5 @@
 class OrdersController < ApplicationController
+  before_action :authenticate_customer!
   def index
     @orders = current_customer.orders.all
   end
@@ -18,46 +19,48 @@ class OrdersController < ApplicationController
   def create
     order = current_customer.orders.new(session[:order])
     order.send_fee = 800
-    order.save
-    if session[:destination] == nil
+    if order.save
+      Destination.exists?(session[:order]["address_tosend"])
       destination = current_customer.destinations.new(
       postcode_tosend: session[:order]["postcode_tosend"],
       address_tosend:  session[:order]["address_tosend"],
       name_tosend:     session[:order]["name_tosend"]
       )
       destination.save
-    else
-    end
 
-    @cart = Cart.where(customer_id: current_customer)
-    @cart.each do |cart|
-      Detail.create(
-      order_id: order.id,
-      product_vol: cart.vol,
-      product_price: (cart.product.price * 1.1).round,
-      production_status:0,
-      product_id: cart.product.id
-      )
+      @cart = Cart.where(customer_id: current_customer)
+      @cart.each do |cart|
+        Detail.create(
+        order_id: order.id,
+        product_vol: cart.vol,
+        product_price: (cart.product.price * 1.1).round,
+        production_status:0,
+        product_id: cart.product.id
+        )
+      end
+      session[:order] = nil
+      session[:destination] = nil
+      current_customer.carts.destroy_all
+      redirect_to homes_thanks_path
+    else
+      @destination = Destination.where(customer_id: current_customer)
+      @order = Order.new
+      @orders = Order.all
+      render :new
     end
-    session[:order] = nil
-    session[:destination] = nil
-    current_customer.carts.destroy_all
-    redirect_to homes_thanks_path
   end
 
   def confirm
     @cart = Cart.where(customer_id: current_customer)
     session[:order] = Order.new()
     if params[:address].to_i == 2
-      session[:order][:pay] = params[:pay].to_i
+      session[:order][:pay] = params[:pay]
       session[:order][:postcode_tosend] = current_customer.postal_code
       session[:order][:address_tosend]  = current_customer.address
       session[:order][:name_tosend]     = current_customer.last_name + current_customer.first_name
       session[:order][:order_status] = 0
     elsif params[:address].to_i == 3
-      destination = current_customer.destinations.find(params[:order][:destination])
-      session[:destination] = params[:order][:destination]
-      session[:order][:pay] = params[:pay].to_i
+      session[:order][:pay] = params[:pay]
       session[:order][:postcode_tosend] = destination.postcode_tosend
       session[:order][:address_tosend]  = destination.address_tosend
       session[:order][:name_tosend]     = destination.name_tosend
@@ -70,7 +73,7 @@ class OrdersController < ApplicationController
       redirect_to orders_path
     end
     subtotal = 0
-      @cart.each do |cart|
+    @cart.each do |cart|
       subtotal += cart.product.price * cart.vol
     end
     session[:order][:total_price] = (subtotal * 1.1).round
